updating utc_date_secs to signed big integer type

+++ src/main/java/com/g414/st9/proto/service/helper/H2TypeHelper.java
@@ -28,7 +28,7 @@ public class H2TypeHelper implements SqlTypeHelper {
         case U64:
             return "BIGINT UNSIGNED";
         case UTC_DATE_SECS:
-            return "INT";
+            return "BIGINT";
         case REFERENCE:
         case UTF8_SMALLSTRING:
             return "VARCHAR(255)";
